rmc/Contrataciones/requirements-2.txt
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.15.0
blinker==1.8.2
Bottleneck==1.3.8
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.17.0
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.3
fonttools==4.51.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
num2words==0.5.13
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
priority==2.0.0
prompt_toolkit==3.0.47
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.31.0
retrying==1.3.4
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
tenacity==8.3.0
tornado==6.4.1
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.1
wcwidth==0.2.13
Werkzeug==3.0.3
wsproto==1.2.0
xlwings==0.31.1
zipp==3.18.1
zope.interface==6.4.post2
